chore: refactor banner seeder
