+++ test/unit/controller/banner-controller.ts
@@ -35,13 +35,13 @@ import User, { TermsOfServiceStatus, UserType } from '../../../src/entity/user/u
 import TokenMiddleware from '../../../src/middleware/token-middleware';
 import RoleManager from '../../../src/rbac/role-manager';
 import Swagger from '../../../src/start/swagger';
-import { seedBanners } from '../../seed-legacy';
 import BannerImage from '../../../src/entity/file/banner-image';
 import { DiskStorage } from '../../../src/files/storage';
 import { defaultPagination, PaginationResult } from '../../../src/helpers/pagination';
 import { truncateAllTables } from '../../setup';
 import { finishTestDB } from '../../helpers/test-helpers';
 import { RbacSeeder } from '../../seed';
+import BannerSeeder from '../../seed/banner';
 
 export function bannerEq(a: Banner, b: BannerResponse): Boolean {
   const aEmpty = a === {} as Banner || a === undefined;
@@ -109,7 +109,7 @@ describe('BannerController', async (): Promise<void> => {
     await User.save(adminUser);
     await User.save(localUser);
 
-    const { banners } = await seedBanners([adminUser, localUser]);
+    const { banners } = await new BannerSeeder().seedBanners([adminUser, localUser]);
 
     // create bearer tokens
     const tokenHandler = new TokenHandler({
